refactor(frontend): rename shared control styles
